feat: rename apiintegration folder
